Terminando Paquete Vista -Se crearon nuevos métodos auxiliares de "vista" en el controlador. -Se cambió un tipo de dato que ocasionaba problemas con la "vista" -Se puede agregar personas a las listas

+++ src/com/calificar_entrenador/modelo/Ciclista.java
@@ -16,7 +16,7 @@ public class Ciclista extends Persona implements Serializable {
     private int tiempo;
     private Entrenador entrenador;
 
-    public Ciclista(String nombre, String cedula, byte edad, int tiempo) {
+    public Ciclista(String nombre, String cedula, int edad, int tiempo) {
         super(nombre, cedula, edad);
         this.tiempo = tiempo;
     }
